Gen AI 2.O/MCP/social-media/requirements.txt
mcp[cli]
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
httpx
